asdfasdfasdfasdfasdf

+++ haiku_baselines/TD4_QR/td4_qr.py
@@ -54,8 +54,9 @@ class TD4_QR(Deteministic_Policy_Gradient_Family):
         
         self.opt_state = self.optimizer.init(self.params)
         
-        self.quantile = (jnp.linspace(0.0,1.0,self.n_support+1)[1:] + jnp.linspace(0.0,1.0,self.n_support+1)[:1])/2.0   # [support]
-        self.quantile = jax.device_put(jnp.expand_dims(self.quantile,axis=(0,1)))                                       # [1 x 1 x support]
+        self.quantile = (jnp.linspace(0.0,1.0,self.n_support+1,dtype=jnp.float32)[1:] + 
+                         jnp.linspace(0.0,1.0,self.n_support+1,dtype=jnp.float32)[:1])/2.0   # [support]
+        self.quantile = jax.device_put(jnp.expand_dims(self.quantile,axis=(0,1))).astype(jnp.float32)  # [1 x 1 x support]
         
         print("----------------------model----------------------")
         print(jax.tree_map(lambda x: x.shape, pre_param))
